# api/openapi.yaml
openapi: 3.0.3
info:
  title: AntiNAT Controller API
  version: 1.0.0-beta
  description: >
    Frozen AntiNAT v1.0-beta management API. All resource modifications use
    ETag/If-Match; missing precondition returns 428, revision mismatch 412,
    and port/state conflict 409. Creates accept a durable Idempotency-Key.
    Error responses use the registry in docs/error-codes.md.
servers:
  - url: /
...
paths:
...
  /api/v1/navigation/categories:
    get:
      tags: [navigation]
      operationId: listNavigationCategories
      summary: List navigation categories in order
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NavigationCategory" }
    post:
      tags: [navigation]
      operationId: createNavigationCategory
      summary: Create a category
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NavigationCategoryCreate" }
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NavigationCategory" }
...
  /api/v1/navigation/items:
    get:
      tags: [navigation]
      operationId: listNavigationItems
      summary: List navigation items in order
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/NavigationItem" }
    post:
      tags: [navigation]
      operationId: createNavigationItem
      summary: Create an item linked to a forward
      security:
        - sessionAuth: []
      parameters:
        - { $ref: "#/components/parameters/IdempotencyKeyHeader" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NavigationItemCreate" }
      responses:
        "201":
          description: Item created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NavigationItem" }
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: antinat_session
  parameters:
...
    IdempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: true
      description: Durable idempotency key for creates
      schema: { type: string, minLength: 8, maxLength: 128 }
...
  schemas:
...
    NavigationCategory:
      type: object
      required: [id, name, order_index, etag]
      properties:
        id: { type: string }
        name: { type: string }
        order_index: { type: integer }
        etag: { type: string }
    NavigationCategoryCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        order_index: { type: integer }
...
    NavigationItem:
      type: object
      required: [id, name, category_id, forward_id, etag]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        protocol: { type: string }
        category_id: { type: string }
        forward_id: { type: string }
        order_index: { type: integer }
        etag: { type: string }
    NavigationItemCreate:
      type: object
      required: [name, category_id, forward_id]
      properties:
        name: { type: string }
        description: { type: string }
        category_id: { type: string }
        forward_id: { type: string }
        order_index: { type: integer }
...